+++ pkg/crd/crd.go
@@ -1,0 +1,12 @@
+package crd
+
+import (
+	_ "embed"
+)
+
+//go:embed charts.yaml
+var charts string
+
+func ChartCRD() string {
+	return charts
+}